Split app up in more modular files
